relocate theme
